+++ Exodus/Changes.md
@@ -1,7 +1,9 @@
 # Utopia Exodus - List of Changes
 
 ## v2.1.1 (in development)
-Supports changes in Citadel
+### Other Changes
+- Supports changes in **Citadel** and **Vault**
+- Deleted some deprecated access objects
 
 ## v2.1 - 3.10.2021
 This update reflects changes in the **Citadel** module, where email address was made optional. This feature / setting